requirements.txt
beautifulsoup4>=4.12
guessit>=3.8
progressbar2>=4.3
requests>=2.32
